+++ src/main/java/com/elmakers/mine/bukkit/plugins/magic/WandChestRunnable.java
@@ -10,24 +10,34 @@ import org.bukkit.scheduler.BukkitRunnable;
 import com.elmakers.mine.bukkit.utilities.NMSUtils;
 
 public class WandChestRunnable extends BukkitRunnable {
-	World world;
-	int dx = 1;
-	int dz = 0;
-	int segmentLength = 1;
-	int x;
-	int z;
-	int segmentPassed = 0;
-	int maxy;
+	private World world;
+	private int dx = 1;
+	private int dz = 0;
+	private int segmentLength = 1;
+	private int x;
+	private int z;
+	private int segmentPassed = 0;
+	private int chunksProcessed = 0;
+	
+	final static int messageInterval = 100;
+	
 	WandChestPopulator populator;
 	Random random;
 	Logger logger;
+	boolean finished = false;
 	
 	public WandChestRunnable(Spells spells, World world, int maxy) {
 		this.world = world;
-		this.maxy = maxy;
 		this.random = new Random();
 		logger = spells.getPlugin().getLogger();
-		populator = spells.getWandChestPopulator();
+		if (maxy > 0) {
+			populator = spells.getWandChestPopulator();
+			populator.setMaxY(maxy);
+		}
+	}
+	
+	public boolean isFinished() {
+		return finished;
 	}
 	
 	public void run() {
@@ -35,11 +45,22 @@ public class WandChestRunnable extends BukkitRunnable {
 		if (!NMSUtils.isDone(chunk) || !chunk.isLoaded()) {
 			if (!NMSUtils.isDone(chunk) || !chunk.load(false)) {
 				logger.info("Done populating chests, found ungenerated chunk");
+				finished = true;
 				this.cancel();
 			}
 		} else {
-			logger.info("Looking for chests in chunk at " + x + "," + z);
-			populator.populate(world, random, chunk);
+			if ((chunksProcessed % messageInterval) == 0) {
+				if (populator != null) {
+					logger.info("Looking for chests, processed " + chunksProcessed + " chunks");
+				} else {
+					logger.info("Looking for wands, searched " + chunksProcessed + " chunks");
+				}
+			}
+			chunksProcessed++;
+			
+			if (populator != null) {
+				populator.populate(world, random, chunk);
+			}
 			x += dx;
 			z += dz;
 			segmentPassed++;